requirements.txt
playwright==1.40.0
aiohttp==3.9.1
aiofiles==23.2.1
python-dotenv==1.0.0
Pillow==10.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
tenacity==8.2.3